[Guerilla] Fix character encoding issues with tool output.
Uses a hacky workaround for Guerilla incorrectly reading the pipe, delaying the read seems to work for now.

+++ H2Codez/H2Sapien.cpp
@@ -363,7 +363,7 @@ void H2SapienPatches::Init()
 
 	GlobalMemoryStatusEx(&statex);
 
-	int SystemMemory = statex.ullAvailPhys / (1024 * 1024);
+	int SystemMemory = static_cast<int>(statex.ullAvailPhys / (1024 * 1024));
 	int VideoMemory = conf.getNumber("VideoMemory", 100) * 1024 * 1024; // megabytes --> bytes
 	int use_hardware_vertexprocessing = conf.getBoolean("use_hardware_vertexprocessing", true);
 	WriteValue(0xF84D1C, SystemMemory);
